Move composition.py Tests into composition.py
We are for now going to keep tests at the bottom of the file where the
test target is implemented.
